Agregando Autorización de acciones con politica en controladores
Se creo la clase ExpenseReportPolicy donde se encuentran las políticas de autorización.
Se añadió esa politica a el proveedor de servicios AuthServiceProvider.
Se agregó validación por recursos y normales a el controlador de ExpenseReport y Expense.

+++ app/Http/Controllers/ExpenseController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Expense;
 use App\ExpenseReport;
 use App\Http\Requests\StoreExpenses;
+use \Illuminate\Auth\Access\AuthorizationException;
 use http\Env\Response;
 use Illuminate\Http\Request;
 
@@ -20,10 +21,13 @@ class ExpenseController extends Controller
      *
      * @param $id
      * @return void
+     * @throws AuthorizationException
      */
     public function create($id)
     {
         $expenseReport = ExpenseReport::findOrfail($id);
+        $this->authorize('view', $expenseReport);
+
         return view('expense.create', [
             'report' => $expenseReport,
         ]);
@@ -35,9 +39,11 @@ class ExpenseController extends Controller
      * @param StoreExpenses $request
      * @param $id
      * @return void
+     * @throws AuthorizationException
      */
     public function store($id, StoreExpenses $request)
     {
+        $this->authorize('view', ExpenseReport::find($id));
         $request->validated();
 
         $expense = new Expense();
@@ -52,15 +58,17 @@ class ExpenseController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $report
-     * @param  int  $expense_id
+     * @param int $report_id
+     * @param int $expense_id
      * @return void
+     * @throws AuthorizationException
      */
-    public function destroy($report, $expense_id)
+    public function destroy($report_id, $expense_id)
     {
+        $this->authorize('delete', ExpenseReport::find($report_id));
         $expense = Expense::findOrfail($expense_id);
         $expense->delete();
 
-        return redirect()->route('expense_reports.show', $report);
+        return redirect()->route('expense_reports.show', $report_id);
     }
 }
